requirement-docker.txt
# Core web framework
Flask==1.1.2
Jinja2==2.10.3 
itsdangerous==1.1.0
MarkupSafe==1.1.1
Werkzeug==0.16.1
Flask-Cors==3.0.10
gunicorn==20.1.0

# Machine learning
tensorflow==1.14.0
Keras==2.2.4
scikit-learn==0.24.1
numpy==1.20.2
pandas==1.1.5
scipy==1.6.2
Pillow==8.2.0
h5py==2.10.0

# Utilities
python-dateutil==2.8.1
pytz==2021.1
requests==2.31.0
python-dotenv==0.19.0